XSI-8 add logging when new domain is detected by RRDD
We suspect that inaccurate a RRD values for VIFs are connected to VIFs
being a "derived" RRD data source that is calculated wrong over a fast
reboot. A derived RRD takes the difference between two counter readings
except when a domain is new - in which case it takes the counter value
as is. We are seeing large values and suspect that they belong to the
counter of the old domain rather than the new domain.

This commit simply adds a log line when a change of domain ID is
detected by the RRD monitor to get a better understanding of the
situation.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>